requirements.txt
torch==1.13.0
transformers==4.23.1
hydra-core==1.1.1
omegaconf==2.1.1
numpy==1.23.4
elasticsearch==7.13.1
sklearn==0.0
tqdm==4.64.1
scipy==1.9.3